+++ Nos/Views/DiscoverGrid.swift
@@ -11,7 +11,7 @@ struct DiscoverGrid: View {
     
     @Environment(\.managedObjectContext) private var viewContext
     @FetchRequest(fetchRequest: Event.emptyDiscoverRequest()) var events: FetchedResults<Event>
-    @ObservedObject var searchModel: SearchModel
+    @ObservedObject var searchController: SearchController
     
     @Binding var columns: Int
     @State private var gridSize: CGSize = .zero {
@@ -25,27 +25,27 @@ struct DiscoverGrid: View {
     
     @Namespace private var animation
 
-    init(predicate: NSPredicate, searchModel: SearchModel, columns: Binding<Int>) {
+    init(predicate: NSPredicate, searchController: SearchController, columns: Binding<Int>) {
         let fetchRequest = Event.emptyDiscoverRequest()
         fetchRequest.predicate = predicate
         fetchRequest.fetchLimit = 1000
         _events = FetchRequest(fetchRequest: fetchRequest)
         _columns = columns
-        self.searchModel = searchModel
+        self.searchController = searchController
     }
     
     var body: some View {
         VStack {
             GeometryReader { geometry in
                 Group {
-                    if searchModel.query.isEmpty {
+                    if searchController.query.isEmpty {
                         StaggeredGrid(list: events, columns: columns) { note in
                             NoteButton(note: note, style: .golden)
                                 .matchedGeometryEffect(id: note.identifier, in: animation)
                         }
                     } else {
                         // Search results
-                        StaggeredGrid(list: searchModel.authorSuggestions, columns: columns) { author in
+                        StaggeredGrid(list: searchController.authorSuggestions, columns: columns) { author in
                             AuthorCard(author: author)
                                 .matchedGeometryEffect(id: author.hexadecimalPublicKey, in: animation)
                         }
